+++ examples/simple/assets/js/app.jsx
@@ -1,7 +1,8 @@
-var React = require('react');
+import React from 'react';
+import '../css/app.css';
 
-module.exports = React.createClass({
-   render: function(){
-       return <h1>Hello, world.</h1>
-   }
-});
+const App = () => {
+  return <h1>Hello, world.</h1>;
+}
+
+export default App;