+++ tests/webpack.config.app2.js
@@ -1,17 +1,14 @@
 var config = require("./webpack.config.simple.js");
-var BundleTracker = require('webpack4-bundle-tracker');
-var MiniCssPlugin = require("mini-css-extract-plugin");
+var BundleTracker = require('webpack-bundle-tracker');
+var MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 config.entry = {
   app2: './assets/js/index'
 };
 
 config.plugins = [
-  new MiniCssPlugin(),
-  new BundleTracker({
-    filename: "./webpack-stats-app2.json",
-    publicPath: "/static/bundles/",
-  }),
+  new MiniCssExtractPlugin(),
+  new BundleTracker({path: __dirname, filename: './webpack-stats-app2.json'})
 ];
 
 module.exports = config;