+++ webpack_loader/__init__.py
@@ -1,4 +1,4 @@
 __author__ = 'Owais Lone'
-__version__ = '0.8.4'
+__version__ = '1.1.0'
 
 default_app_config = 'webpack_loader.apps.WebpackLoaderConfig'